# Use numa_sched_setaffinity to bind thread to each CPU core