--- a/.github/workflows/ccpp.yml
+++ b/.github/workflows/ccpp.yml
@@ -26,7 +26,7 @@ jobs:
         run: brew install boost
       - name: install dependencies (Linux)
         if: matrix.os == 'ubuntu-latest'
-        run: sudo apt update && sudo apt install -yq libboost-dev g++-7
+        run: sudo apt update && sudo apt install -yq libboost-dev g++-7 libnuma-dev
       - name: checkout submodules
         shell: bash
         if: matrix.os != 'macos-latest' || matrix.compiler != 'g++-7'