--- a/src/thread_pool/thread_pool.cpp
+++ b/src/thread_pool/thread_pool.cpp
@@ -56,7 +56,6 @@ ThreadPool::ThreadPool(size_t pool_size)
       }
     });
   }
-  // TODO NEED COUNTDOWN LATCH
 }
 
 ThreadPool::~ThreadPool() {