This PR uses numa_sched_setaffinity to bind thread to each CPU core, where “thread” is every worker thread in the LineairDB’s inner thread pool.
It suspects reducing cache misses after context switches, preemption, or yield-loop for waiting transactions.

### benchmark result

w/o numa_sched_setaffinity (master, f2d6a47 )
```
bench/ycsb -t 72 -q 36 => 264557 tps
bench/ycsb -t 72 -q 18 => 236214 tps
bench/ycsb -t 36 -q 36 => 323508 tps
bench/ycsb -t 36 -q 18 => 281465 tps
bench/ycsb -t 18 -q 8  => 450423 tps
```

w/ numa_sched_setaffinity
```
bench/ycsb -t 72 -q 36 => 235136 tps
bench/ycsb -t 72 -q 18 => 229287 tps
bench/ycsb -t 36 -q 36 => 356228 tps
bench/ycsb -t 36 -q 18 => 339929 tps
bench/ycsb -t 18 -q 8  => 635739 tps
```

The number of physical cores is 72 in these experiments.
Further details on the environment are described in [README.md](https://github.com/LineairDB/LineairDB/blob/master/README.md).
We can see 10%~40% performance gain under settings that spawns threads ≦ the number of physical cores.
At the settings higher than the number of physical cores (where hyper-threading comes into working in this case), we can see a slight (but negligible) decrease in TPS. I'm still investigating this issue but I consider that it is caused by the cache misses when multiple workers run together on the same physical core.

This PR resolves #238.